requirements.txt
requests>=2.31.0
python-slugify>=8.0.0
openai>=1.0.0

